requirements.txt
flask>=2.0.0
langchain>=0.0.300
langchain-community>=0.0.300  # Add this line
groq>=0.1.0
python-dotenv>=0.19.0
requests>=2.26.0
langchain_groq>=0.0.300
mysql-connector-python>=8.0.26
fastapi>=0.115.11
uvicorn>=0.34
httpx>=0.20.0